+++ CHANGELOG.md
@@ -8,3 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Pages: Catalog, Contact, Login, Sign up, Product Details, Manage Products.
+
+### Fixed
+
+- Navbar height.